+++ BusanIT_JSP_Project/src/login/LoginBean.java
@@ -1,0 +1,50 @@
+package login;
+
+public class LoginBean {
+	
+	private String id;
+	private String pwd;
+	private String name;
+	private String phonenumber;
+	private String cPostNumber;
+	private String cAddress;
+	
+	public String getId() {
+		return id;
+	}
+	public void setId(String id) {
+		this.id = id;
+	}
+	public String getPwd() {
+		return pwd;
+	}
+	public void setPwd(String pwd) {
+		this.pwd = pwd;
+	}
+	public String getName() {
+		return name;
+	}
+	public void setName(String name) {
+		this.name = name;
+	}
+	public String getPhonenumber() {
+		return phonenumber;
+	}
+	public void setPhonenumber(String phonenumber) {
+		this.phonenumber = phonenumber;
+	}
+	public String getcPostNumber() {
+		return cPostNumber;
+	}
+	public void setcPostNumber(String cPostNumber) {
+		this.cPostNumber = cPostNumber;
+	}
+	public String getcAddress() {
+		return cAddress;
+	}
+	public void setcAddress(String cAddress) {
+		this.cAddress = cAddress;
+	}
+	
+		
+}
